Support for Maxmind's GeoIP2 library.
This removes support for their (and our) legacy product.  It also removes
the feature that downloaded the free databases from the web if you
requested an update.  This is because Maxmind no longer as a free
version of the ASN/ISP data for GeoIP2 and thus the feature didn't
really make complete sense.